// server/api/spotify/types.ts
export type TrackItem = Track | Episode;
// ...
export interface Page<TItemType> {
    href: string;
    items: TItemType[];
    limit: number;
    next: string | null;
    offset: number;
    previous: string | null;
    total: number;
}

export interface PlaylistedTrack<Item extends TrackItem = TrackItem> {
    added_at: string;
    added_by: AddedBy;
    is_local: boolean;
    primary_color: string;
    track: Item;
}

export interface AddedBy {
    external_urls: ExternalUrls;
    href: string;
    id: string;
    type: string;
    uri: string;
}
// ...
export interface Track extends SimplifiedTrack {
    album: SimplifiedAlbum;
    external_ids: ExternalIds;
    popularity: number;
}
// ...
export interface Followers {
    href: string | null;
    total: number;
}

export interface ExternalUrls {
    spotify: string;
}
// ...
export interface Image {
    url: string;
    height: number;
    width: number;
}
// ...
export interface Episode extends SimplifiedEpisode {
    show: SimplifiedShow;
}
// ...
interface PlaylistBase {
    collaborative: boolean;
    description: string;
    external_urls: ExternalUrls;
    followers: Followers;
    href: string;
    id: string;
    images: Image[];
    name: string;
    owner: UserReference;
    primary_color: string;
    public: boolean;
    snapshot_id: string;
    type: string;
    uri: string;
}

export interface Playlist<Item extends TrackItem = TrackItem>
    extends PlaylistBase {
    tracks: Page<PlaylistedTrack<Item>>;
}

export interface FeaturedPlaylists {
    message: string;
    playlists: Page<SimplifiedPlaylist>;
}

export interface SimplifiedPlaylist extends PlaylistBase {
    tracks: TrackReference | null;
}

export interface TrackReference {
    href: string;
    total: number;
}

export interface UserReference {
    display_name: string;
    external_urls: ExternalUrls;
    href: string;
    id: string;
    type: string;
    uri: string;
}